requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
basicsr==1.4.2
beautifulsoup4==4.12.2
blendmodes==2022
blinker==1.6.2
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.4
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
einops==0.4.1
exceptiongroup==1.1.2
facexlib==0.3.0
fastapi==0.94.0
ffmpy==0.3.0
filelock==3.12.2
filterpy==1.4.5
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
gevent==23.7.0
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.32.0
gradio_client==0.2.9
greenlet==2.0.2
grpcio==1.56.0
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
inflection==0.5.1
Jinja2==3.1.2
jsonmerge==1.8.0
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
kornia==0.6.7
lark==1.1.2
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
llvmlite==0.40.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
open-clip-torch @ https://github.com/mlfoundations/open_clip/archive/bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b.zip
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==2.0.3
piexif==1.1.3
Pillow==9.5.0
platformdirs==3.8.1
protobuf==3.20.0
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.3.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
safetensors==0.3.1
scikit-image==0.20.0
scipy==1.11.1
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
streamlit==1.24.1
sympy==1.12
tb-nightly==2.14.0a20230712
tenacity==8.2.2
tensorboard-data-server==0.7.1
tifffile==2023.7.10
timm==0.6.7
tokenizers==0.13.3
tomesd==0.1.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchdiffeq==0.2.3
torchmetrics==1.0.0
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
trampoline==0.1.2
transformers==4.25.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
yapf==0.40.1
yarl==1.9.2
zipp==3.16.1
zope.event==5.0
zope.interface==6.0
